backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
argon2-cffi==23.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyjwt==2.8.0
email_validator==2.3.0
fastapi-sso==0.17.0
asyncpg==0.29.0
redis==6.4.0
celery==5.3.4
prometheus-client==0.20.0
structlog==23.2.0
elasticsearch==8.13.0
python-json-logger==2.0.7
sentry-sdk>=1.40.0
hawk-python-sdk[fastapi]